fix: #61 build payment intent only when basket is truthy
Requests to make Stripe payment intent with a null 'amount' value in the view would crash the page. The payment intent is now only created when the basket has a value. Furthermore, the script that instantiates the Stripe elements is only included when the basket is truthy within the template context.